+++ lib/dam/endpoints.go
@@ -80,6 +80,10 @@ const (
 	// Part of SCIM V2 for managing groups. See "proto/scim/v2/groups.proto"
 	scimGroupsPath = "/identity/scim/v2/{realm}/Groups"
 	scimGroupPath  = "/identity/scim/v2/{realm}/Groups/{name}"
+	// Part of SCIM V2 for managing users. See "proto/scim/v2/users.proto"
+	scimUsersPath = "/identity/scim/v2/{realm}/Users"
+	scimUserPath  = "/identity/scim/v2/{realm}/Users/{name}"
+	scimMePath    = "/identity/scim/v2/{realm}/Me"
 
 	// testPath: performs a set of tests for validity of the current configuration.
 	// TODO: remove and perform tests at the time of config update and reject